+++ ConsoleCloneCOLUMNS/ConsoleCloneCOLUMNS/BaseStage.cpp
@@ -1,0 +1,11 @@
+#include "BaseStage.h"
+
+
+BaseStage::BaseStage()
+{
+}
+
+
+BaseStage::~BaseStage()
+{
+}