+++ ConsoleCloneCOLUMNS/ConsoleCloneCOLUMNS/Columns.h
@@ -1,0 +1,19 @@
+#pragma once
+#include "Common.h"
+#include "ControlStage.h"
+#include "ViewStage.h"
+#include "Jewel.h"
+class Columns
+{
+private:
+	ControlStage* controlStage;
+	ViewStage* viewStage;
+	Jewel* jewel;
+public:
+	Columns();
+	~Columns();
+
+	void init();
+	void draw();
+};
+