+++ ConsoleCloneCOLUMNS/ConsoleCloneCOLUMNS/Common.h
@@ -1,0 +1,31 @@
+#pragma once 
+#include <iostream>
+#include <time.h>
+//#define DEBUG
+
+enum DEFAULTSTAGESIZE
+{
+	defaultStageWidth = 6,
+	defaultStageHeight = 13,
+	defaultViewStageWidth = 8,
+	defaultViewStageHeight = 17,
+};
+
+enum OBJECT
+{
+	SPACE,
+	RED,
+	YELLOW,
+	BLUE,
+	GREEN,
+	ORANGE,
+	PERPLE,
+
+	BLOCK = 99,
+};
+
+enum JEWELDARA{
+
+	defDewelNum = 3,
+	defJewelNum = 6,
+};