+++ HISTORY.md
@@ -1,8 +1,14 @@
 # HISTORY
-##0.1.1(2014/03/25)
+##0.1.0(2014/03/25)
  ### 最初のリリース
  * LICENSE.mdの追加
  * README.mdの追加
  * HISTORY.mdの追加
- 
- 
+ ##0.2.0(2014/03/26)
+ ### 各クラス作成
+ * ステージ表示の追加
+ * 
+ 備考<br>
+ Jewel,ControlStage,ViewStageクラスの
+ データを元にColumnsで作成していこうか、
+ ControlStageにJewelに持っていって色々するか考え中。